requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
qrcode[pil]>=7.4.2
Pillow>=9.5.0
aiosqlite>=0.19.0
psutil>=5.9.0  # NEW: For memory monitoring
